reformat: create seperate file to calculate grid
